msp430_nrf24 - configure to work as transmitter, sending pressure and temp.  Hardware includes the BMP280 pressure/temp ic.  removed uart debug output and added led green to 2.2

+++ source/ccs/msp430_nrf24/sensors/BMP280.c
@@ -85,14 +85,10 @@ void BMP280_init(void)
 	//result - 0001 0000 = 0x10
 	BMP280_writeReg(BMP280_REG_CONFIG, 0x10);
 
-	//try standby 125ms - 010  100 00
-//	BMP280_writeReg(BMP280_REG_CONFIG, 0x50);
-
-
 	//wakeup - bits 0 and 1 high for reg BMP280_REG_CTRL_MEAS
 	BMP280_writeReg(BMP280_REG_CTRL_MEAS, 0x57);
 
-	//Load Calibration data into ram - loads mCalibrationData
+	//Load Calibration data into ram - mCalibrationData
 	BMP280_readCalibrationValues();
 
 }
@@ -143,25 +139,19 @@ BMP280_Data BMP280_read(void)
 	xlsb = ((uint32_t)(data[5] >> 4)) & 0x0F;
 	uTemperature = ((msb << 12) | (lsb << 4) | (xlsb));
 
-//	result.uPressure = uPressure;
-//	result.uTemperature = uTemperature;
-
-	//compute the corrected values using the raw data
-	//and calibration values as described in the datasheet
-
-	//compute the compensated temp
+	//correct values using calibration data
 	BMP280_S32_t cTemperature = BMP280_compensate_T_int32((BMP280_S32_t)uTemperature);	//div 100 to get int, % 100 for frac
 	BMP280_U32_t cPressure = BMP280_compensate_P_int64((BMP280_S32_t)uPressure);		//divide 256 for integer
 
 	//integer and fractional - celcius
-	result.cTemperatureCInt = (uint16_t)(cTemperature / 100);
-	result.cTemperatureCFrac = (uint16_t)(cTemperature % 100);
+	result.cTemperatureCInt = (uint8_t)(cTemperature / 100);
+	result.cTemperatureCFrac = (uint8_t)(cTemperature % 100);
 
 	BMP280_S32_t cTemperatureF = cTemperature * 9 / 5;
 	cTemperatureF += 3200;
 
-	result.cTemperatureFInt = (uint16_t)(cTemperatureF / 100);
-	result.cTemperatureFFrac = (uint16_t)(cTemperatureF % 100);
+	result.cTemperatureFInt = (uint8_t)(cTemperatureF / 100);
+	result.cTemperatureFFrac = (uint8_t)(cTemperatureF % 100);
 
 	result.cPressure = cPressure / 256;				//pressure in pa
